Updated the filter icon logic to work with prototype as well
