+++ app/bundles/LeadBundle/Controller/AjaxController.php
@@ -26,6 +26,7 @@ use Mautic\LeadBundle\Provider\FormAdjustmentsProviderInterface;
 use Mautic\LeadBundle\Segment\Stat\SegmentCampaignShare;
 use Mautic\LeadBundle\Services\ContactColumnsDictionary;
 use Mautic\LeadBundle\Services\SegmentDependencyTreeFactory;
+use Mautic\LeadBundle\Twig\Extension\SegmentFilterIconTrait;
 use Mautic\PluginBundle\Helper\IntegrationHelper;
 use Symfony\Component\Form\FormFactoryInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -35,6 +36,7 @@ use Symfony\Component\HttpFoundation\Response;
 class AjaxController extends CommonAjaxController
 {
     use AjaxLookupControllerTrait;
+    use SegmentFilterIconTrait;
 
     public function userListAction(Request $request): JsonResponse
     {
@@ -882,8 +884,8 @@ class AjaxController extends CommonAjaxController
     public function updateLeadFieldOrderChoiceListAction(Request $request): Response
     {
         $object = InputHelper::clean($request->request->get('object'));
-        $group  = InputHelper::clean($request->request->get('group'));
-        $field  = new LeadField();
+        $group = InputHelper::clean($request->request->get('group'));
+        $field = new LeadField();
         $field->setObject($object);
         $field->setGroup($group);
         $form = $this->createForm(FieldType::class, $field);
@@ -894,5 +896,12 @@ class AjaxController extends CommonAjaxController
                 'form' => $form->createView(),
             ]
         );
+    }
+
+    public function getSegmentFilterIconAction(Request $request): JsonResponse
+    {
+        $fieldAlias = $request->request->get('filterType');
+
+        return $this->sendJsonResponse(['iconClass' => $this->getSegmentFilterIcon($fieldAlias)]);
     }
 }
